Struct rt_easy_unit_test::unit_test::Value

pub struct Value { /* fields omitted */ }

Implementations

impl Value

pub fn zero(size: usize) -> Value

pub fn one(size: usize) -> Value

pub fn filled(size: usize) -> Value

pub fn remove_leading_zeros(&mut self)

pub fn extend_zero(&mut self, size: usize)

pub fn extend_sign(&mut self, size: usize)

pub fn with_size(self, size: usize) -> Value

pub fn parse_bin(bin: &str) -> Result<Value, ()>

Parse from binary string. Leading zeros from input are retained.

Errors

Errors if bin is empty or contains any other char then [01].

pub fn parse_dec(dec: &str) -> Result<Value, ()>

Parse from decimal string. The result will have no leading zeros.

Errors

Errors if dec is empty or contains any other char then [0-9].

pub fn parse_hex(hex: &str) -> Result<Value, ()>

Parse from hexadecimal string. The result will have no leading zeros.

Errors

Errors if hex is empty or contains any other char then [0-9a-fA-F].

pub fn concat<'a, I>(slices: I) -> Value where
I: IntoIterator<Item = &'a ValueSlice>,
<I as IntoIterator>::IntoIter: DoubleEndedIterator,

pub fn as_slice(&self) -> &ValueSlice

pub fn as_mut_slice(&mut self) -> &mut ValueSlice